Add lots of consts.
In general, we should use 'const fz_blah' in device calls whenever
the callee should not alter the fz_blah.

Push this through. This shows up various places where we fz_keep
and fz_drop these const things.

I've updated the fz_keep and fz_drops with appropriate casts
to remove the consts. We may need to do the union dance to avoid
the consts for some compilers, but will only do that if required.

I think this is nicer overall, even allowing for the const<->no const
problems.