Force all harfbuzz allocations through our allocators.
Because of a shortcoming in harfbuzz, we can't easily force
all its allocations through our allocators.

We fudge it, with the addition of some macros to change
malloc/free/calloc into hb_malloc/hb_free/hb_calloc. To
prevent thread safety issues, we use our freetype lock
around calls to harfbuzz.

We stash the current context in a static var.