PCL output: Add code to guess paper sizes.
Guess paper sizes from bitmap dimensions unless the paper
size has been specified. If unknown, depending on the
properties of the printer, we'll either send a 'custom'
paper size (followed by the page size in decipoints), or
the next size up.

Depending on the capabilities of the printer, we can
optionally send orientation too.