Clarify scissor argument to clip device functions.
The scissor argument is an optional (potentially NULL) rectangle
that can give hints to devices about the area that can be scissored.
This is used by the draw device and display list device to minimize
the size of temporary clip mask buffers.

The scissor rectangle, if used, must have been transformed by the
current transform matrix.