Strip outlines when cleaning a pdf file.
Arrange to keep only the Outline entries that still refer to
valid pages.

More specifically, leaf outline entries that refer to pages that
are elided are not dropped. Non leaf outline entries that refer
to pages that have been elided are kept (for the sake of the
children) but have NULL destinations.