Rejig Bidirectional and Text code.
We move to using bidirectional "levels" throughout. This should
give us better behaviour vis-a-vis nested l2r/l2r text. This
also allows us to carry xps levels throughout with no loss of
information.

This also avoids the need to special case numbers.

We accordingly carry more information into fz_text.

As well as wmode, also hold additional details about the text spans.

We now include the directionality of the bidi level text (either as
derived from bidi code, or from the original document (e.g. xps)),
the directionality of text (as specified in the original document
(e.g. html)), and the language of the text (if specified in the
original document).