Bug 696668: Update the downscaling logic.
An l2factor of 3 is equivalent to downscaling by a factor of 8.

We can get an l2factor of 3 downscale out of the jpeglib. We can
reasonably downscale by a further l2factor of 3 manually. Any more
than that and we start to completely drop pixels without them
having any effect.

Therefore it's pointless us keeping any tiles around with l2factors
> 6.

Fix the bug (which was that we were using < instead of <=) and
update the value to a more reasonable one anyway.