More ARM code thresholding work.
Fix do_threshold_1 and implement do_threshold_4.

do_threshold_1 checks for white and shortcuts the work if it
can.

There are 2 ARM variants of do_threshold_4. One for ARMs that
support unaligned loads, and one for the (rare configurations)
that don't. The former checks for white and shortcuts the work.