Memento: Add facility to cope with 'known' leaked blocks.
Harfbuzz allocates blocks that it keeps around as statics.
There are a finite number of these, and they are never freed
(at least not in the build we use). Having Memento report
them as leaks is a pain.

So we extend Memento so that we can label blocks as being
leaks, and thus never have to be bothered by them.