Add 'Low Memory' mode to Mutool draw.
If used, PDF objects are cleared at the end of each page, and
the store size is set to 1 byte, avoiding the caching of images.

This trades repeated decoding of images/reading of file objects
for memory usage. The purpose of this is to enable measurements to
be made on the minimum possible memory level.