Fix corruption of file using sanitize.
When sanitizing a file, while cleaning with decompression, I was
seeing a flate problem reported.

The issue is that pdf_open_filter was passing pdf_open_raw_filter
the orig_num as both num and orig_num. This was causing us to
find an fz_buffer attached to the (wrong) xref entry and to open
that instead of the underlying stream.

The fix is to propogate num a bit further.