Fix broken documents after sanitize
The DP and BDC operators, are used in the form:

 <NAME> <PROPERTIES> <OPERATOR>

where <PROPERTIES> can either be a name (that can be looked
up to get a dictionary) or an inline dictionary.

What the spec doesn't say is that the two are not
interchangeable. If you resolve the name to an inline
dict, then insert it, Acrobat will give an error for
some (but not all) cases.

The interpreter currently resolves any references, and
passes the resolved version into the operator handling
function. This precludes us outputting the original
form.

We therefore update it to pass both the raw and the cooked
versions in. This has no effect on MuPDFs own handling of
anything, it just enables the buffer device to output
a correct stream.