Tweak pdf-write option handling.
The handling of not-decompressing images/fonts was geared towards
pdfclean usage; but now that we can create new PDF files, it makes
more sense to ask for images and fonts to be compressed, rather than
asking for them not to be decompressed with quirky interaction with
the 'expand' and 'deflate' flags.

If -f or -i are set, we will never decompress images, and we will
compress them if they are uncompressed.

If -d is set, we will first decompress all streams (module -f or -i).
If -z is set, we will then compress all uncompressed streams.