Bug 696795: Proper fix.
When affine plotting with linear interpolation, we need to perform
different calculations for the texture and for the edge of the shape.

The edge of the shape needs to be calculated in exactly the same way
as for non-linear interpolated shapes.

The 'texture' position needs to be offset by 1/2 a texture unit in
each direction so that the 'pure' color is given in the middle of
the texture cells image, not in the top left corner.

To achieve these aims, we actually offset the u/v positions by 1/2
(32768, given the fixed point we are using) and adjust for this in
the boundary tests.

I have a test file that shows this working, which I will attach to
the bug, and add to the regression suite.