stext: Non-initial glyphs in ligatures must set start/stop.
Normal glyphs and inital glyphs in ligatures have their start/stop (p
and q) set before determining whether to append to an existing span or
insert a space. For non-initial glyphs the start/stop were never set
which introduced uninitialized values into the span data structure.

Now, all glyphs have their start/stop set and then if it is a
non-initial glyph in a ligature the append and space detection is
ignored. This means that no values are uninitialized.