Update test device with passthrough option.
Update the test device so it can be used 'wrapping' another
device.

In particular, it can be used to wrap the display list device
so that we can evaluate 'color or not' while building the
display list rather than having to rerun the display list
afterwards.

Also, give improved control over whether we test every pixel
of images/shadings.