Bug 696822: Refuse to save a pdf file incrementally if it would break.
If a file cannot be saved incrementally, then don't accept that
as an option. In practise this means if someone asks to save
a file incrementally, and it was repaired, or it uses encryption
then throw an error.

Add a new function to ask if it's safe to save a file incrementally,
and use that in the appropriate places.