Update TIFF LZW decode.
TIFF 5.0 uses a slightly laxer set of rules for TIFF decode.

Specifically, when we hit the maximum code, we are not required
to send a clear code immediately, but it can overrrun.

We don't bother storing codes > 12 bits, because they can never
be used. This avoids the need to extend the table.