Fix inconsistency in band handling.
In an earlier commit, I changed some of the band writing functions
to take the band starting offset, rather than the band number. This
was done to accomodate the idea of rendering the page in bands of
different heights.

Sadly, it seems I didn't push this all the way through, and had
different band writing functions still taking the band number.

Fix all the band writing functions to be consistent.