Fix garbage collection and page grafting for indirect reference chains.
The mark & sweep pass of garbage collection, and resolving indirect objects when grafting objects
was following the full chain of indirect references. In the unusual case where a numbered object
is itself only an indirect reference to another object, this intermediate numbered object would
be missed both when marking for garbage collection, and when copying objects for grafting.

Add a function to resolve only one step for these two uses.

The following is an example of a file that would break during garbage collection if we
follow full indirect reference chains:

%PDF-1.3

1 0 obj
  <</Type/Catalog /Foo[2 0 R 3 0 R]>>
endobj

2 0 obj
  4 0 R
endobj

3 0 obj
  5 0 R
endobj

4 0 obj
  <</Length 1>>
stream
A
endstream
endobj

5 0 obj
  <</Length 1>>
stream
B
endstream
endobj