Android JNI Annotation fix.
We pass pointers from the JNI layer into the java to be stored
as java longs. When doing this it is VITAL that we cast them
correctly as otherwise the JNI layer on android has problems -
presumably because of alignment in the procedure calling
standard.

I'd missed this in one place. Fixed here.

<log></log>