Bug 695988: Avoid sending image data multiple times.
If the same image is used repeatedly, then we currently send
the data again and again. Instead, send each image as a
reusable symbol and reuse it as required.

This reduces the size of the output by over 50% for the example
file, but unfortunately results in the example file killing
both Firefox and Edge when we attempt to load it. The file
seems entirely valid though, and this is a saving on sane files.

There is a #define that can be made to revert to the old
behaviour (SVG_SEND_REPEATED_IMAGES).