zip: Used unpacked entry size, not listed entry size.
The size listed in the central directory can be used to allocate an
output buffer and inflate can be told to not write more than this
number of bytes. The listed size cannot be assumed to be the size
of the output data however as the zip file may be corrupted. Instead
the size of the output data must be given by inflate and must be
less than or equal to the listed size.

The same reasoning goes for uncompressed entries and simply reading
the uncompressed data from the archive file (which may terminate
early).

Fixes indeterminism broken_png_image.xps.