Correct rounding in gproof skeleton files.
In the gproof pipeline MuPDF first generates a skeleton file
with page sizes in. Then it gets gs to generate individual
pages for that file.

In one case, the pdf file has a height of 203.4 - at 300dpi
that's 847.5 device pixels. MuPDF was rounding that up to
848pixels, whereas gs was rounding it down to 847, causing
an error on load.

Adjust MuPDFs skeleton file generation so that it matches
gs.