More tweaking of gproof device gs invocation.
If USE_GS_API is defined, we use the gsapi interface to call
gs.

The gsapi interface is defined in iapi.h, which is not included
in the mupdf source code. We therefore include a potted version
of this API inline.

To use the original, pass HAVE_IAPI_H at compile time.