Bug 697151: Pacify valgrind when reading padding while unpacking tiles
If e.g. an 18 x 11 PNG image used 1 bit depth and a single component
the source data fed to fz_unpack_tile() would have padding bits at the
end of each line of whole bytes, because 18 bits is not a multiple of 8.

The optimized 1 component versions of fz_unpack_tile() for a bit depth
of 1 deliberately read the padding when doing the table lookups, knowing
that this does not matter because of the construction of the table.
Valgrind is incapable of understanding this, so we provide some extra
masking (used only in PACIFY_VALGRIND builds) that explicitly mask
off the invalid bits.