Bug 697177: Avoid SEGV in Epub file.
Thanks to Vlad Glagolev for the example file, analysis and
patch.

We had failed to initialise the "last break class" value, so that
when we use it as an index into an array, we run the risk of
crashing.

The fix is simply to initialise it. By initialising it to OP
we don't allow any softbreaks to be added when it's uninitialised,
which feels right to me.