Optimise fz_new_buffer_from_image_as_png.
This function calls png_from_pixmap, an auxiliary function to do
the work. This was done deliberately to allow us to make the
optimisation of dropping the pixmap early if it is not required.
This keeps the memory use down in the case when a color conversion
is required.

Sadly, I never appear to have made the trivial change to actually
use it in this mode. Do that now.