Bug 697234: Fix slow rendering.
fz_contains_rect was improperly implemented (the logic in the
final test was reversed). This was causing us to use the font
bbox for some t3 glyphs, resulting in massive bboxes, that could
never be cached. Rerendering these each time was taking ages,
even though there was nothing actually in them.