Fix cluster timeouts with test file.
tests_private/pdf/sumatra/1297_-_interpolate_at_lower_resolutions.pdf
times out in the cluster.

This is due to us having empty t3 glyphs defined that define d1
rectangles that are wildly different to the default font bbox.

Add code to spot that t3 glyphs are empty, and to use a tiny font
bbox for them.

(It might be nicer to drop the empty display lists, but then this
produces knock on problems further on, where non-existent display
lists lead to NULL pixmaps, which lead to us think that renders
failed etc).