Bug 697226: Fix SEGV in Android viewer.
As we skip through pages very quickly, it is apparently possible
to trigger a SEGV. Alex Talis has given a clear description of
the problem on the bug, and proposed this solution.

Essentially this tweaks our CancellableAsyncTask class to ensure
that we do not destroy the cookie before it has finished being
accessed.