Bug 697241: Fix blending through clips.
Non rectangular clips are currently handled by rendering to a
'isolated' background, and then plotting that through a mask.

This runs into problems when the rendering needs to use non
standard blend modes that need to access the background
colors.

Instead, copy the background to the new pixmap, render to that
then plot that through the mask.

This simplifies the painting code, because we now never have
mismatched source and destination alphas.