Add simple layer configuration feature to mutool draw.
This commit adds a new -y flag to mutool draw. -y takes an
additional argument - this can either be a single 'l' meaning
"list the layer configs available", or a comma separated list
of numbers.

The first number in the list says which "layer config" to
select. Subsequent numbers identify lines in the config to
toggle.

For instance I might do:

  mutool draw -y l -o out.png layers.pdf

and this will list me the layer configs in the file (as well
as rendering normally).

From this I might choose layer config 0:

 mutool draw -y 0 -o out.png layers.pdf

This will select layer config 0, and show a table (max 1 layer
per numbered line) that shows which layers are on or off.

If I then decide that I want layers 3 and 4 to be toggled from
their default state, I can run:

 mutool draw -y 0,3,4 -o out.png layers.pdf

Short of implementing an interactive configuration mode, this
is the simplest way to demonstrate the layer functionality.