Continued Layer API work.
Hide pdf_is_hidden_ocg from the public interface.

Ensure that RBGroups and Order properly fall back to the default
entries if they are missing in the OCCDs.

Add pdf_set_layer_config_as_default. This sets the default
layer config to be current one. This destroys the current
default dictionary, but we leave the other configs intact.