Fix Java 32/64bit detection.
Fix some java typos. Thanks to Fred for nicely pointing out that
I'd been testing the wrong build.

Also, make us default to 32bit if sun.arch.data.model is
not defined, as Android doesn't set this.

We will need to find a nice way to spot 32/64 bittedness on
Android.