Bug 697301: Fix "crash" in epub.
Actually an assert. This is caused by a paragraph separator in the
text. The Unicode Bidirectional Algorithm says we should operate
paragraph by paragraph, and includes code to split paragraphs
at paragraph markers, changing their type to boundary neutrals
as it goes.

The use of this code was left "as an exercise for the reader" in
the example code, so we simply hook it up here.