Tweak pdf portfolio: automatically use incremental writing.
Windows cannot remove open files, so we can't save non-incrementally
to the same file name as the open document. Force incremental saves
if the input and output file names are the same.

NOTE: We will also create corrupt files if saving a new document incrementally,
so take care not to turn on incremental saves if we have created a new blank
document.