Ensure we don't clobber files in fz_new_output_with_path.
When saving a PDF to the same file name as the one that is open,
we risk clobbering it by truncating the source file. Remove the old
file first! Removing an open file will not work on windows,
but there we will throw an EACCESS error rather than clobber the file!