Avoid using .incbin when compiling for Android.
The clang 3.8 used in Android NDK r13 supports .incbin, but doesn't
take any -I arguments into account when looking for the file to be
included. In Android we depend on -I to locate the font files to be
included. For ARM-based targets this works becased Android NDK
supplies -fno-integrated-as to cause the compiler to use an external
assembler that searches the -I arguments to locate the file to
include. For x86-based targets -fno-integrated-as is not supplied
so the files cannot be located and thus the compilation fails.

To avoid this problem and be as robust as possible let's disable the
.incbin directive for Android for all architectures and compilers.