epub: Adjust y coordinate for page_top in draw calls, not by top matrix.
We used to bake in the page_top in all the y coordinates of all objects,
and use the top level matrix to subtract page_top.

If we subtract page_top from the y coordinates instead of using the
matrix to do it, we get smaller numbers in the device calls.