Fix MSVC solution.
If we have mutool depend on libluratech, then it won't build
when luratech is not installed.

If we don't have mutool depend on libluratech, then it won't
build the Commercial versions.

The basic problem is that dependencies in MSVC can't be dependent
on configuration.

Therefore, move the guts of mutool into libmutool, and have mutool
and mutool-lura projects that depend on it. mutool-lura can also
depend on libluratech.

mutool builds in non commercial configurations, and mutool-lura in
the commercial ones.