Fix needs_reap entry in key storable.
There is no need to hold a separate flag to say that
we need reaping, when this can be implied from the
store_key_refs and the normal refcount being equal.

In addition, I don't think we were ever actually
setting this, so the code was wrong to start with.