Rendering intent support for graphic and text
Graphic fills and strokes as well as text fill and strokes
handle the rendering intent settings.  This works
through the display list.  The parameters related to
color rendering, ri, bp, op, opm are packed in the flags
bits through the display list.

To do: Add support for images and shadings.