Add clamping proc to colorspace-imp.h
In the head of mupdf, the index color values were not
being clamped properly.  I moved the clamping operation
to be performed by a procedure in of the color space.

The a*, b* values in LAB will need to include a range value
for proper clamping as well as the ICC color space at some
point.