Add front end support for default color spaces
PDF can specify in the resource dictionary color spaces
that should be used for DeviceGray, DeviceRGB and DeviceCMYK.
This commit handles the extraction of these color spaces
and passes them through the display list and gets them
into a structure on a draw device (if one is used).
Next step is to have the draw device make use of them.