Backend of default color spaces
This handles the use of page level definitions for
DefaultGray, DefaultRGB, DefaultCMYK in the
draw device.   The interface for the pixmap color
conversion had to be expanded a bit to pass the
default_cs object.  This was needed due to the fact
that the image could be DeviceN or Separation and
have a base space that needs to be replaced by
a page level Default color space.   Tested this
with a file that had two pages each with an common
reference image object but different DefaultCMYK
definitions in the page resource dictionary.  Proper
rendering was obtained.