Add icc support to band-writer and png output
For image viewers that support icc profiles, it
is important that we include the icc profile
into those formats that support.  Currently the
only output format we have that supports it is
PNG.  But if we decide to do tiff, jpeg or psd
those also support the embedding of icc profiles.