jni: Clear java exception when rethrow via fz_throw().
Previously when a java exception was pending CallObjectMethod()
was called. This is not permitted and so the exception must be
cleared before this function is called.

Secondly if the method called by CallObjectMethod() throws an
exception this exception was left pending when fz_throw() for
unknown java errors was called. Later on jni_rethrow() would
be called which in turn called ThrowNew(). This is not
permitted either, so the pending exception must be cleared
if caused by the call to CallObjectMethod().

Because fz_throw_java() always ends with a fz_throw() it is
safe to assume that the pre-existing exception (or the one
caused by CallObjectMethod()) will always be converted into
a fz_throw(), which in turn will be converted back to a java
exception since all calls to fz_throw_java() are enclosed in
fz_try() where fz_catch() ends with calling jni_rethrow().